+++ ndsampler/coco_frames.py
@@ -6,6 +6,56 @@ from os.path import join
 import ubelt as ub
 
 
+if 0:
+    # Maybe?
+    import smqtk
+    import abc
+    smqtk.Pluggable = smqtk.utils.plugin.Pluggable
+    smqtk.Configurable = smqtk.utils.configuration.Configurable
+
+    """
+    {
+        "frames": {
+            "ndsampler.CogFrames:0": {
+                'compression': 'JPEG',
+            },
+            "ndsampler.NPYFrames": {
+            },
+            "type": "ndsampler.CogFrames"
+        }
+    }
+    """
+
+    class AbstractFrames(smqtk.Pluggable, smqtk.Configurable):
+
+        def __init__(self, config, foo=1):
+            pass
+
+        @abc.abstractmethod
+        def load_region(self, spec):
+            pass
+
+        @classmethod
+        def is_usable(cls):
+            return True
+
+    class CogFrames(AbstractFrames):
+
+        # default_config = {
+        #     'compression': Value(
+        #         default='JPEG', choices=['JPEG', 'DEFLATE']),
+        # }
+        # def __init__(self, **kwargs):
+        #     super().__init__(**kwargs)
+
+        def load_region(self, spec):
+            return spec
+
+    class NPYFrames(AbstractFrames):
+        def load_region(self, spec):
+            return spec
+
+
 class CocoFrames(abstract_frames.Frames, util.HashIdentifiable):
     """
     wrapper around coco-style dataset to allow for getitem syntax
@@ -36,6 +86,7 @@ class CocoFrames(abstract_frames.Frames, util.HashIdentifiable):
                                          workdir=workdir, backend=backend)
         self.dset = dset
         self.verbose = verbose
+        self._image_ids = None
 
     def _lookup_gpath(self, image_id):
         img = self.dset.imgs[image_id]
@@ -45,9 +96,13 @@ class CocoFrames(abstract_frames.Frames, util.HashIdentifiable):
             gpath = img['file_name']
         return gpath
 
-    @ub.memoize_property
+    @property
     def image_ids(self):
-        return list(self.dset.imgs.keys())
+        if self._image_ids is None:
+            import numpy as np
+            # Use ndarrays to prevent copy-on-write as best as possible
+            self._image_ids = np.array(list(self.dset.imgs.keys()))
+        return self._image_ids
 
     def _make_hashid(self):
         _hashid = getattr(self.dset, 'hashid', None)